Save coords to the db
With this commit I implemented the functionality to save coords received from the client to the db and print the result to the screen to check that everything went well. It actually isn't going too well, since db entries are stored in a weird format at the moment. I will look into this lately.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>