Replace mammoccio with user to avoid confusion
The whole mammoccio thing was getting confused, so now when we are talking about the logged in user, we call it user, while when we are talking about mammoccio, we are talking about a Mammoccio instance stored in the database. This should avoid confusion going forward (see the next commit, for example).

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>